split examples into different executables, some code organization
