+++ src/common/avx_mathfun_tweaked.h
